implementado restante dos metodos crud so serviceImpl

+++ src/main/java/br/com/fiap/livraria/repository/LivroRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface LivroRepository extends JpaRepository<Livro, Integer> {
+public interface LivroRepository extends JpaRepository<Livro, Long> {
 
 
     @Query("from Livro l " +
